feat(disorders): adiciona paginação nuqs e aplica vertical slicing (#837)
